requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
reportlab==4.2.5
requests==2.32.3
sqlparse==0.5.1
sslcommerz-lib==1.0
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
